requirements.txt
# CEO Data Agent Demo - Python Dependencies

# Data Generation
pandas>=2.0.0
numpy>=1.24.0

# Azure SDK
azure-identity>=1.14.0
azure-storage-blob>=12.19.0
azure-keyvault-secrets>=4.7.0

# OpenAI Integration
openai>=1.12.0

# Semantic Kernel (for advanced orchestration)
semantic-kernel>=0.9.0

# Fabric Integration
msal>=1.26.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
python-dotenv>=1.0.0
ipykernel>=6.25.0
